Create .gitignore
Things seem to be reading into the html object fine. It has a length of 30 (the number of urls opened by the "u in url" for loop. However, what I want in the "s in soup" loop is for it to grab each class element of the soup object and append it as a separate item in the info list object. So, while soup would have 30 objects, info would have 300. It isn't working, though--info has a length of 10, and when I run the "i in info" for loop, I get a "'ResultSet' object has no attribute 'text'" error and the obs list object isn't populated. Any tips on what I'm doing wrong?